Began rework of Service Callbacks
