+++ daemon/src/services/garmin/garminmessages.h
@@ -10,11 +10,15 @@
 #include <QSet>
 #include <QDateTime>
 #include <QtGlobal>
+#include <QSharedPointer>
 #include <optional>
 
 #include "garmintypes.h"
-#include <variant>
+//#include "communicator_v2.h"
 
+// Forward declarations
+class CommunicatorV2;
+//class ServiceCallback;
 
 
 
@@ -125,145 +129,26 @@ inline std::optional<QString> messageIdToString(quint16 id) {
 }
 
 
-/// Status codes for response messages
-///
-enum class Status : quint8 {
-    Ack = 0,
-    Nack = 1,
-    Unsupported = 2,
-    DecodeError = 3,
-    CrcError = 4,
-    LengthError = 5,
-};
-
-inline std::optional<Status> statusFromU8(quint8 code) {
-    switch (code) {
-    case 0: return Status::Ack;
-    case 1: return Status::Nack;
-    case 2: return Status::Unsupported;
-    case 3: return Status::DecodeError;
-    case 4: return Status::CrcError;
-    case 5: return Status::LengthError;
-    default: return std::nullopt;
-    }
-}
-
-inline QString statusName(Status s) {
-    switch (s) {
-    case Status::Ack: return "ACK";
-    case Status::Nack: return "NACK";
-    case Status::Unsupported: return "UNSUPPORTED";
-    case Status::DecodeError: return "DECODE_ERROR";
-    case Status::CrcError: return "CRC_ERROR";
-    case Status::LengthError: return "LENGTH_ERROR";
-    }
-    return "UNKNOWN";
-}
-
-// -------------------- Parsed message structs --------------------
-// Device Information Message (incoming from watch)
-
-
-struct DeviceInformationMessage {
-    quint16 protocolVersion{};
-    quint16 productNumber{};
-    quint32 unitNumber{};
-    quint16 softwareVersion{};
-    quint16 maxPacketSize{};
-    QString bluetoothFriendlyName;
-    QString deviceName;
-    QString deviceModel;
-};
-
-// Configuration Message (incoming from watch)
-
-struct ConfigurationMessage {
-    QSet<quint16> capabilities;
-};
-
-// Notification Control Message (incoming from watch)
-
-struct NotificationControlMessage {
-    qint32 notificationId{};
-    quint8 command{};
-    QVector<QPair<quint8, quint16>> attributes; // (attribute_id, max_length)
-    std::optional<quint8> actionId;
-    std::optional<QString> actionString;
-};
-
-// Notification Subscription Message (incoming from watch)
-
-struct NotificationSubscriptionMessage {
-    bool enable{};
-    quint8 unk{};
-};
-
-// Synchronization Message (incoming from watch)
-
-struct SynchronizationMessage {
-    quint8 synchronizationType{};
-    quint64 fileTypeBitmask{};
-
-    bool shouldProceed() const {
-        constexpr quint64 WORKOUTS = 1ull << 3;
-        constexpr quint64 ACTIVITIES = 1ull << 5;
-        constexpr quint64 ACTIVITY_SUMMARY = 1ull << 21;
-        constexpr quint64 SLEEP = 1ull << 26;
-        return (fileTypeBitmask & (WORKOUTS | ACTIVITIES | ACTIVITY_SUMMARY | SLEEP)) != 0;
-    }
-};
-
-// Filter Status Message (response to Filter message)
-
-struct FilterStatusMessage {
-    Status status{};
-    quint8 filterType{};
-};
-
-// Weather Request Message (5014)
-
-struct WeatherRequestMessage {
-    quint8 format{};
-    qint32 latitude{};
-    qint32 longitude{};
-    quint8 hoursOfForecast{};
-};
-
-
-
-struct UnknownMessage {
-    quint16 messageId {0};
-    QByteArray data;
-};
-
-
-using GfdiMessage = std::variant<
-    DeviceInformationMessage,
-    ConfigurationMessage,
-    std::monostate, // CurrentTimeRequest
-    NotificationControlMessage,
-    NotificationSubscriptionMessage,
-    SynchronizationMessage,
-    FilterStatusMessage,
-    WeatherRequestMessage,
-    UnknownMessage
->;
-
 
 // -------------------- Parser (QObject + signals) --------------------
 
-class GfdiMessageParser : public QObject {
+class GfdiMessageParser : public ServiceCallback {
     Q_OBJECT
 public:
-    explicit GfdiMessageParser(QObject* parent=nullptr) : QObject(parent) {}
+    static QSharedPointer<GfdiMessageParser> create(CommunicatorV2* parent=nullptr) {
+        return QSharedPointer<GfdiMessageParser>(new GfdiMessageParser(parent));
+    }
+    explicit GfdiMessageParser(CommunicatorV2* parent=nullptr) : mCommunicator(parent)
+    {
+    }
+    void setCommunicator(CommunicatorV2* comm);
+    void onMessage(const QByteArray& data) override;
+
+    void parse(const QByteArray& data);
 
 public slots:
     // Convenience slot: parse and emit appropriate signal
     //void parseAndEmit(const QByteArray& data);
-
-public:
-
-    void parse(const QByteArray& data);
 
 signals:
     void deviceInformationReceived(DeviceInformationMessage& msg);
@@ -289,6 +174,8 @@ private:
 
     static Result<QString> readLengthPrefixedString(const QByteArray& data, int& consumedBytes);
     static QSet<quint16> parseCapabilities(const QByteArray& bytes);
+
+    CommunicatorV2* mCommunicator;
 
 
 };
